Separate extra commands
